waitingに移動, propsをvotingとgameに反映
