+++ client/src/components/user.tsx
@@ -1,0 +1,5 @@
+
+
+export type User = {
+    user: string|undefined;
+}
