+++ client/src/pages/game.tsx
@@ -1,11 +1,20 @@
 import UserList from "../components/userlist"
+import {User} from '../components/user'
 
 const Game = () => {
+  const users: User[] = [
+    {
+      user: 'A'
+    },
+    {
+      user: 'B'
+    }
+  ];
     return (
       <div>
         <h1>Game</h1>
         <p>This is the game page</p>
-        <UserList users={['A','B']}/>
+        <UserList users={users}/>
       </div>
     )
   }
